make node ids integers when consolidating intersections
When consolidating intersections, weakly connected components in the
same cluster are separated and given different cluster labels. The
cluster label is the original cluster label, plus a prefix for the
component, separated by a '-'. Since osm xml format requires integer
osmids, the clusters need to be renamed after components are separated.