Use Session for HTTP requests
Include a default requests.Session as part of settings that is used for HTTP requests to OSM APIs, and can be overridden using the utils.config. This should improve performance by re-using the underlying TCP connection when sending multiple requests to the same host, and allows a user of osmnx to override the session to provide client/server certificates when making request to any OSM API that uses 2-way SSL verification.