Refactor to resolve server closing connection
Diagnosing the intermittent test failures, determined that the root cause was the server closing the connection due to inactivity. Refactored the settings to expose a dictionary of attributes that can be configured for a requests.Session that is instantiated when a call is made to _osm_network_download, _osm_geometries_download, or nominatum_request. Less elegant, but should allow users to provide authentication and 2-way SSL configuration, while also improving performance during repeated calls to the overpass API.