add add_node_elevations_opentopodata function
hence there are APIs between google and opentopodata are quite similar,
they share the same underlying function.

for opentopodata it is possible to customize the endpoint because it
can be hosted locally. by default it uses the public API.
the dataset is customizeable as well because the user might want use a
different one. the default is aster30m, which works worldwide to 30m