Only convert datatypes to str after removing null values
In the current code, the gdf_nodes and gdf_edges dataframes are converted to strings early on. This effectively renders calls to dropna() useless. For exemple, float NaNs are converted to 'nan', and thus dropna() does nothing. This commit keeps the correct datatypes so that we can call dropna() later on (i.e. on a per-row basis) and only convert to string when creating XML elements.