Update graph.py
Similar functionality is available for the function graph_from_point(). But it is not available for def graph_from_polygon(). Here user can get buffer with fixed 500 meters. However the value of buffer is not flexible. My proposal is that we can also add one new parameter buffer_dist(float value) so that user can give any buffer value to create a graph with the buffer.